pyproject.toml
[tool.poetry]
name = "polygon-api"
version = "0.1.0"
description = ""
authors = ["Eric Barrow <eric.barrow@pnmac.com>"]

packages = [
    { include = "common", from = "src/core"},
    { include = "processor", from = "src"},
    { include = "processor_service", from = "src/processor"}
]

[tool.poetry.dependencies]
python = "^3.8"
polygon-api-client = "^1.1.0"
python-dotenv = "^0.20.0"
fastapi = "^0.78.0"
mangum = "^0.15.0"
marshmallow = "^3.17.0"
aws-lambda-powertools = "^1.27.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
aws-cdk-lib = "^2.13.0"
boto3 = "^1.24.36"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
